Fix storybook to provide a github token.
