+++ .env.template
@@ -1,0 +1,1 @@
+STORYBOOK_GITHUB_TOKEN=${GITHUB_TOKEN_READONLY}
