--- .storybook/preview.ts
@@ -1,15 +1,0 @@
-import type { Preview } from "@storybook/react";
-import './global.css';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
